Adds secret santa functionality to lists
Adds columns to lists and secret santa rounds tables to support secret santa functionality.

Specifically, it adds a flag to the lists table to indicate whether a list is for secret santa, and it adds columns to the secret santa rounds table for signup cutoff date, auto-draw enable, push notification enable, and email notification enable.